+++ dds/src/implementation/dds_impl/mod.rs
@@ -1,18 +1,18 @@
 pub mod any_data_reader_listener;
 pub mod any_data_writer_listener;
 pub mod any_topic_listener;
-pub mod builtin_data_reader;
+pub mod node_builtin_data_reader;
 pub mod builtin_publisher;
 pub mod builtin_stateful_reader;
 pub mod builtin_stateful_writer;
 pub mod builtin_stateless_reader;
 pub mod builtin_stateless_writer;
-pub mod builtin_subscriber_impl;
-pub mod builtin_subscriber_node;
+pub mod builtin_subscriber;
+pub mod node_builtin_subscriber;
 pub mod dcps_service;
 pub mod domain_participant_impl;
-pub mod listener_data_reader;
-pub mod listener_data_writer;
+pub mod node_listener_data_reader;
+pub mod node_listener_data_writer;
 pub mod listener_subscriber;
 pub mod message_receiver;
 pub mod node_kind;
@@ -21,12 +21,12 @@ pub mod reader_factory;
 pub mod status_condition_impl;
 pub mod status_listener;
 pub mod topic_impl;
-pub mod user_defined_data_reader_node;
+pub mod node_user_defined_data_reader;
 pub mod user_defined_data_reader;
-pub mod user_defined_data_writer_node;
+pub mod node_user_defined_data_writer;
 pub mod user_defined_data_writer;
 pub mod user_defined_publisher;
-pub mod user_defined_publisher_node;
+pub mod node_user_defined_publisher;
 pub mod user_defined_subscriber;
-pub mod user_defined_subscriber_node;
+pub mod node_user_defined_subscriber;
 pub mod writer_factory;
