+++ dds/src/implementation/dds_impl/node_builtin_subscriber.rs
@@ -5,7 +5,7 @@ use crate::{
 };
 
 use super::{
-    builtin_subscriber_impl::{BuiltInSubscriber, BuiltinDataReaderKind},
+    builtin_subscriber::{BuiltInSubscriber, BuiltinDataReaderKind},
     domain_participant_impl::DomainParticipantImpl,
 };
