+++ dds/src/implementation/dds_impl/node_kind.rs
@@ -1,9 +1,9 @@
 use super::{
-    builtin_data_reader::BuiltinDataReaderNode, builtin_subscriber_node::BuiltinSubscriberNode,
-    listener_data_reader::ListenerDataReaderNode, listener_data_writer::ListenerDataWriterNode,
-    listener_subscriber::ListenerSubscriberNode, user_defined_data_reader_node::UserDefinedDataReaderNode,
-    user_defined_data_writer_node::UserDefinedDataWriterNode,
-    user_defined_subscriber_node::UserDefinedSubscriberNode,
+    node_builtin_data_reader::BuiltinDataReaderNode, node_builtin_subscriber::BuiltinSubscriberNode,
+    node_listener_data_reader::ListenerDataReaderNode, node_listener_data_writer::ListenerDataWriterNode,
+    listener_subscriber::ListenerSubscriberNode, node_user_defined_data_reader::UserDefinedDataReaderNode,
+    node_user_defined_data_writer::UserDefinedDataWriterNode,
+    node_user_defined_subscriber::UserDefinedSubscriberNode,
 };
 
 #[derive(PartialEq, Debug)]
