+++ dds/src/implementation/dds_impl/node_user_defined_data_writer.rs
@@ -21,7 +21,7 @@ use super::{
     any_data_writer_listener::AnyDataWriterListener,
     domain_participant_impl::DomainParticipantImpl, status_condition_impl::StatusConditionImpl,
     topic_impl::TopicImpl, user_defined_data_writer::UserDefinedDataWriter,
-    user_defined_publisher_node::UserDefinedPublisherNode,
+    node_user_defined_publisher::UserDefinedPublisherNode,
     user_defined_publisher::UserDefinedPublisher,
 };
