+++ dds/src/implementation/dds_impl/user_defined_data_writer.rs
@@ -53,7 +53,7 @@ use crate::{
 
 use super::{
     any_data_writer_listener::AnyDataWriterListener, domain_participant_impl::AnnounceKind,
-    listener_data_writer::ListenerDataWriterNode, message_receiver::MessageReceiver,
+    node_listener_data_writer::ListenerDataWriterNode, message_receiver::MessageReceiver,
     status_condition_impl::StatusConditionImpl, status_listener::StatusListener,
     topic_impl::TopicImpl, user_defined_publisher::UserDefinedPublisher,
 };
